+++ test.sh
@@ -40,7 +40,7 @@ limit_test=${1:-}
 declare -A EXPECTED
 EXPECTED['./sacrebleu.py -t wmt16,wmt17 -l en-fi --echo ref | ./sacrebleu.py -b -w 4 -t wmt16/B,wmt17/B -l en-fi']=53.7432
 EXPECTED['./sacrebleu.py -t wmt16,wmt17 -l en-fi --echo ref | ./sacrebleu.py -b -w 4 -t wmt16/B,wmt17/B -l en-fi --origlang=en']=18.9054
-EXPECTED['./sacrebleu.py -t wmt17 -l en-fi --echo ref | ./sacrebleu.py -b -t wmt17/B -l en-fi --verbose']="55.6
+EXPECTED['./sacrebleu.py -t wmt17 -l en-fi --echo ref | ./sacrebleu.py -b -t wmt17/B -l en-fi --detail']="55.6
 origlang=en : N=1502 BLEU= 21.4
 origlang=fi : N=1500 BLEU=100.0"